mongoIntegration

+++ tpPOIs-3052-2-domain/src/main/java/dds/poi/builder/AsyncProcessResultBuilder.java
@@ -18,7 +18,7 @@ public class AsyncProcessResultBuilder extends Builder<AsyncProcessResult>{
 		return this;
 	}
 	
-	public AsyncProcessResultBuilder userId(int userId) {
+	public AsyncProcessResultBuilder userId(Long userId) {
 		this.asyncProcessResult.setUserId(userId);
 		return this;
 	}
